+++ backend/src/app.module.ts
@@ -1,14 +1,11 @@
 import { Module } from '@nestjs/common';
-import { ConfigModule } from '@nestjs/config';
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+import { DescriptionModule } from './modules/description/description.module';
 
 @Module({
-  imports: [
-    ConfigModule.forRoot({
-      isGlobal: true,
-      envFilePath: '.env',
-    }),
-  ],
-  controllers: [],
-  providers: [],
+  imports: [DescriptionModule],
+  controllers: [AppController],
+  providers: [AppService],
 })
-export class AppModule {} 
+export class AppModule {}